refactor: reduce ChoiceSelector.tsx to 278 lines (from 512) and document async change
Extracted logic into focused modules to meet 300-line project standard:
- Created src/lib/utils/requirementNormalizer.ts for item requirement normalization
- Created src/components/shared/ChoiceSelector/optionNormalizer.ts for option processing
- Created src/components/shared/ChoiceSelector/RequirementBadges.tsx for badge rendering
- Created src/components/shared/ChoiceSelector/choiceStyling.tsx for alignment/weight styling

Added JSDoc to narrativeGenerator.ts explaining why formatResponse() became async:
supports AI-based item extraction fallback to ensure inventory-gated decisions have
reliable data.

All 24 ChoiceSelector tests passing.

Related to PR #763 code review recommendations.
